+++ NEWS.md
@@ -1,3 +1,22 @@
+# gsignal 0.3-3
+
+- date: 20220330
+- Fixed Github Issue #3: Problems with fftfilt when FFT length is provided by user
+- copy attributes of input object x to output in functions filter, filtfilt, sosfilt, fftfilt
+- copy dimnames of input object x to output in functions upfirdn, resample, upsample, upsamplefill,
+       downsample, decimate, detrend, fht, sgolayfilt, 
+- added ultrwin() function
+- adapted filter() to allow data and filter coefficients to be of type complex
+- adapted sosfilt() to allow data and filter coefficients to be of type complex
+- bugfix in pwelch() for multivariate input
+- Fixed Github Issue #4: Problem with hilbert for small amplitude signals
+- added isConjSymm() function to gsignal-internal
+- adapted ifft() to use isConjSymm instead of ZapIm
+- reduced default tolerance for isWhole() and zapIm()
+- bugfix in detrend(): function now returns a vector if input was a vector
+- bugfix in filtfilt(): corrected bug in computing filter ends (default and Sos methods)
+
+
 # gsignal 0.3-2
 
 - date: 20210518